useRegisterAsItemViewer hook
